Add youtube dataset named 0301.zip.
    Add read and generate generalization hierarchies.
    Add read data.

TODO:

    related_ID GH.

+++ utils/utility.py
@@ -11,7 +11,7 @@ def cmp_str(element1, element2):
     compare number in str format correctley
     """
     try:
-        return cmp(int(element1), int(element2))
+        return cmp(float(element1), float(element2))
     except ValueError:
         return cmp(element1, element2)
